Solution to Toll Plaza Management System

+++ Toll Plaza Management System@@67877d61dc107486ad5c2750/code.c
@@ -12,7 +12,7 @@ int main() {
     int len;
     scanf("%d", &len);
     struct Vehicle vehicle[len];
-    float car,truck,bike;
+    int car,truck,bike;
     for (int i = 0; i < len; i++) {
         scanf("%s %s %d", &vehicle[i].id, vehicle[i].name, &vehicle[i].toll);
         if (vehicle[i].name == "Car") car+=vehicle[i].toll;
@@ -21,7 +21,7 @@ int main() {
     }
 
 
-     printf("Cars: %.2f, Trucks: %.2f, Bikes: %.2f",car,truck,bike);
+     printf("Cars: %.2f, Trucks: %.2f, Bikes: %.2f",(float)car,(float)truck,(float)bike);
 
 
     return 0;
